Scraping functionality essentially done. Now testing out different models for how analytics should be performed on team and player trends vs different bet types etc.
