+++ package.json
@@ -7,7 +7,7 @@
     "test": "node ./src/test.js",
     "testdb": "node ./src/testSetup.js",
     "run":  "node ./src/index.js",
-    "setup": "source .rc && node ./src/setup.js"
+    "setup": "node ./src/setup.js"
   },
   "author": "",
   "license": "ISC",
