--- src/backend/trends.js
@@ -1,6 +1,0 @@
-// File that defines functions for analyzing team & player trends. This will contains basically all functionality this
-// application is supposed to provide to the user.
-
-function goalsForPeriodAverage(games, period=1) {
-
-}
