+++ src/util/utilities.js
@@ -1,5 +1,4 @@
-const Constants = require('./constants')
-const dateStringify = (date) => date.toISOString().split("T")[0];
+const Constants = require('./constants');
 const l = msg => console.log(msg);
 const getGameURL = (date) => {
     if(date !== String)
@@ -9,6 +8,16 @@ const getGameURL = (date) => {
 const Int = (istr) => Number.parseInt(istr);
 const Float = (fstr) => Number.parseFloat(fstr);
 
+const dateString = (date) => {
+    let res = date.toLocaleDateString();
+    let restmp = res.split("-");
+    let y = Number.parseInt(restmp[0]);
+    let mNum = Number.parseInt(restmp[1]);
+    let dNum = Number.parseInt(restmp[2]);
+    return `${y}-${(mNum < 10) ? `0${mNum}` : mNum}-${(dNum < 10) ? `0${dNum}` : dNum}`
+};
+const dateStringify = (date) => dateString(date);
+
 const removePrefixOf = (str, len) => {
     try {
         if(str.length < len) {
@@ -17,6 +26,23 @@ const removePrefixOf = (str, len) => {
         return str.split("").splice(str.length-4, 4).join("");
     } catch(err) {
         return str;
+    }
+};
+
+const type = (obj) => {
+    let [def, name, ..._] = obj.constructor.toString();
+    if(def === "class") {
+        return name;
+    } else if(def === "function") {
+        return name;
+    }
+};
+
+function dumpErrorStackTrace(err) {
+    let e = new Error("Fuck");
+    if(err) {
+        console.log(`\nError Message: ${err.message}`);
+        console.log(`Stack: ${err.stack}`)
     }
 }
 
@@ -42,14 +68,7 @@ module.exports = {
     getGameURL,
     Int,
     Float,
-    removePrefixOf
+    removePrefixOf,
+    dateString
 };
 
-const type = (obj) => {
-    let [def, name, ..._] = obj.constructor.toString();
-    if(def === "class") {
-        return name;
-    } else if(def === "function") {
-        return name;
-    }
-}
